requirements.txt
git+https://github.com/xivo-pbx/xivo-dao.git
alembic
pyhamcrest
sqlalchemy==0.9.4
